ADMINDASH/requirements.txt
Flask==3.1.2
Flask-CORS==6.0.2
PyJWT==2.11.0
mysql-connector-python==9.4.0
